ARIA_Assistant/requirements.txt
# ── ЯДРО ───────────────────────────────────────────────
kivy==2.3.0
kivymd==1.2.0
Pillow>=10.0.0

# ── ИИ (Groq — бесплатно 14 400 req/день) ──────────────
# Groq использует стандартный urllib (встроен), доп. пакет не нужен
# Опционально для более удобного клиента:
# groq

# ── GOOGLE SHEETS ───────────────────────────────────────
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0

# ── ГОЛОС (Desktop) ─────────────────────────────────────
openai-whisper          # STT (локально, офлайн)
sounddevice             # Запись с микрофона
soundfile               # Аудио файлы
numpy                   # Обработка аудио
pyttsx3                 # TTS (офлайн)

# ── ANDROID ─────────────────────────────────────────────
plyer                   # Кросс-платформа (TTS, уведомления)
# pyjnius               # Устанавливается автоматически через buildozer

# ── УТИЛИТЫ ─────────────────────────────────────────────
schedule                # Планировщик (напоминания)
